refactor: Phase 5 [5.3] migrate losses/ — move criterion/*.py to losses/, add deprecation shims (BPR)
