+++ src/criterion/FocalLoss.py
@@ -1,44 +1,14 @@
-import torch
-import torch.nn.functional as F
-from torch import nn
+"""Shim — canonical location is losses.focal (Phase 5).
 
+This module will be removed at the end of Phase 6.
+"""
+import warnings as _warnings
 
-class FocalLoss(nn.Module):
-    def __init__(self, alpha=None, gamma=2.0, reduction='mean'):
-        """
-        Focal Loss with optional per-class alpha weights.
+_warnings.warn(
+    "criterion.FocalLoss is deprecated; use losses.focal instead. "
+    "This shim will be removed at the end of Phase 6.",
+    DeprecationWarning,
+    stacklevel=2,
+)
 
-        Args:
-            alpha: Class weights. Can be:
-                   - None: no class weighting
-                   - float/int: scalar weight for all classes
-                   - Tensor: per-class weights indexed by target class
-            gamma: Focusing parameter (default=2.0). Higher values focus more on hard examples.
-            reduction: 'mean', 'sum', or 'none'
-        """
-        super(FocalLoss, self).__init__()
-        self.alpha = alpha
-        self.gamma = gamma
-        self.reduction = reduction
-
-    def forward(self, inputs, targets):
-        ce_loss = F.cross_entropy(inputs, targets, reduction='none')
-        pt = torch.exp(-ce_loss)
-        focal_weight = (1 - pt) ** self.gamma
-
-        # Handle per-class alpha weights
-        if self.alpha is not None:
-            if isinstance(self.alpha, (float, int)):
-                alpha_t = self.alpha
-            else:
-                # Index alpha by target class for each sample
-                alpha_t = self.alpha[targets]
-            focal_weight = alpha_t * focal_weight
-
-        loss = focal_weight * ce_loss
-
-        if self.reduction == 'mean':
-            return loss.mean()
-        elif self.reduction == 'sum':
-            return loss.sum()
-        return loss
+from losses.focal import FocalLoss  # noqa: F401
