+++ src/criterion/__init__.py
@@ -1,0 +1,4 @@
+"""Shim package — canonical location is losses/ (Phase 5).
+
+All modules will be removed at the end of Phase 6.
+"""
