+++ src/losses/pcgrad.py
@@ -1,0 +1,184 @@
+import copy
+import random
+from typing import Dict, List, Tuple, Union
+from abc import abstractmethod
+
+import torch
+
+
+class WeightMethod:
+    def __init__(self, n_tasks: int, device: torch.device):
+        super().__init__()
+        self.n_tasks = n_tasks
+        self.device = device
+
+    @abstractmethod
+    def get_weighted_loss(
+            self,
+            losses: torch.Tensor,
+            shared_parameters: Union[List[torch.nn.parameter.Parameter], torch.Tensor],
+            task_specific_parameters: Union[
+                List[torch.nn.parameter.Parameter], torch.Tensor
+            ],
+            last_shared_parameters: Union[List[torch.nn.parameter.Parameter], torch.Tensor],
+            representation: Union[torch.nn.parameter.Parameter, torch.Tensor],
+            **kwargs,
+    ):
+        pass
+
+    def backward(
+            self,
+            losses: torch.Tensor,
+            shared_parameters: Union[
+                List[torch.nn.parameter.Parameter], torch.Tensor
+            ] = None,
+            task_specific_parameters: Union[
+                List[torch.nn.parameter.Parameter], torch.Tensor
+            ] = None,
+            last_shared_parameters: Union[
+                List[torch.nn.parameter.Parameter], torch.Tensor
+            ] = None,
+            representation: Union[List[torch.nn.parameter.Parameter], torch.Tensor] = None,
+            **kwargs,
+    ) -> Tuple[Union[torch.Tensor, None], Union[dict, None]]:
+        """
+
+        Parameters
+        ----------
+        losses :
+        shared_parameters :
+        task_specific_parameters :
+        last_shared_parameters : parameters of last shared layer/block
+        representation : shared representation
+        kwargs :
+
+        Returns
+        -------
+        Loss, extra outputs
+        """
+        loss, extra_outputs = self.get_weighted_loss(
+            losses=losses,
+            shared_parameters=shared_parameters,
+            task_specific_parameters=task_specific_parameters,
+            last_shared_parameters=last_shared_parameters,
+            representation=representation,
+            **kwargs,
+        )
+        loss.backward()
+        return loss, extra_outputs
+
+    def __call__(
+            self,
+            losses: torch.Tensor,
+            shared_parameters: Union[
+                List[torch.nn.parameter.Parameter], torch.Tensor
+            ] = None,
+            task_specific_parameters: Union[
+                List[torch.nn.parameter.Parameter], torch.Tensor
+            ] = None,
+            **kwargs,
+    ):
+        return self.backward(
+            losses=losses,
+            shared_parameters=shared_parameters,
+            task_specific_parameters=task_specific_parameters,
+            **kwargs,
+        )
+
+    def parameters(self) -> List[torch.Tensor]:
+        """return learnable parameters"""
+        return []
+
+
+class PCGrad(WeightMethod):
+    """Modification of: https://github.com/WeiChengTseng/Pytorch-PCGrad/blob/master/pcgrad.py
+
+    @misc{Pytorch-PCGrad,
+      author = {Wei-Cheng Tseng},
+      title = {WeiChengTseng/Pytorch-PCGrad},
+      url = {https://github.com/WeiChengTseng/Pytorch-PCGrad.git},
+      year = {2020}
+    }
+
+    """
+
+    def __init__(self, n_tasks: int, device: torch.device, reduction="sum"):
+        super().__init__(n_tasks, device=device)
+        assert reduction in ["mean", "sum"]
+        self.reduction = reduction
+
+    def get_weighted_loss(
+            self,
+            losses: torch.Tensor,
+            shared_parameters: Union[
+                List[torch.nn.parameter.Parameter], torch.Tensor
+            ] = None,
+            task_specific_parameters: Union[
+                List[torch.nn.parameter.Parameter], torch.Tensor
+            ] = None,
+            **kwargs,
+    ):
+        raise NotImplementedError
+
+    def _set_pc_grads(self, losses, shared_parameters, task_specific_parameters=None):
+        # shared part
+        shared_grads = []
+        for l in losses:
+            shared_grads.append(
+                torch.autograd.grad(l, shared_parameters, retain_graph=True, allow_unused=True)
+            )
+
+        if isinstance(shared_parameters, torch.Tensor):
+            shared_parameters = [shared_parameters]
+        non_conflict_shared_grads = self._project_conflicting(shared_grads)
+        for p, g in zip(shared_parameters, non_conflict_shared_grads):
+            p.grad = g
+
+        # task specific part
+        if task_specific_parameters is not None:
+            task_specific_grads = torch.autograd.grad(
+                losses.sum(), task_specific_parameters
+            )
+            if isinstance(task_specific_parameters, torch.Tensor):
+                task_specific_parameters = [task_specific_parameters]
+            for p, g in zip(task_specific_parameters, task_specific_grads):
+                p.grad = g
+
+    def _project_conflicting(self, grads: List[Tuple[torch.Tensor]]):
+        pc_grad = copy.deepcopy(grads)
+        for g_i in pc_grad:
+            random.shuffle(grads)
+            for g_j in grads:
+                g_i_g_j = sum(
+                    [
+                        torch.dot(torch.flatten(grad_i), torch.flatten(grad_j))
+                        for grad_i, grad_j in zip(g_i, g_j)
+                    ]
+                )
+                if g_i_g_j < 0:
+                    g_j_norm_square = (
+                            torch.norm(torch.cat([torch.flatten(g) for g in g_j])) ** 2
+                    )
+                    for grad_i, grad_j in zip(g_i, g_j):
+                        grad_i -= g_i_g_j * grad_j / g_j_norm_square
+
+        merged_grad = [sum(g) for g in zip(*pc_grad)]
+        if self.reduction == "mean":
+            merged_grad = [g / self.n_tasks for g in merged_grad]
+
+        return merged_grad
+
+    def backward(
+            self,
+            losses: torch.Tensor,
+            parameters: Union[List[torch.nn.parameter.Parameter], torch.Tensor] = None,
+            shared_parameters: Union[
+                List[torch.nn.parameter.Parameter], torch.Tensor
+            ] = None,
+            task_specific_parameters: Union[
+                List[torch.nn.parameter.Parameter], torch.Tensor
+            ] = None,
+            **kwargs,
+    ):
+        self._set_pc_grads(losses, shared_parameters, task_specific_parameters)
+        return None, {}  # NOTE: to align with all other weight methods
